moved LHCInfoPerLS from PPS to RunInfo
