+++ CondCore/CTPPSPlugins/src/plugin.cc
@@ -27,8 +27,6 @@
 #include "CondFormats/DataRecord/interface/PPSAlignmentConfigurationRcd.h"
 #include "CondFormats/PPSObjects/interface/PPSAssociationCuts.h"
 #include "CondFormats/DataRecord/interface/PPSAssociationCutsRcd.h"
-#include "CondFormats/PPSObjects/interface/LHCInfoPerLS.h"
-#include "CondFormats/DataRecord/interface/LHCInfoPerLSRcd.h"
 
 namespace {
   struct InitAssociationCuts {
@@ -50,6 +48,5 @@ REGISTER_PLUGIN(PPSDirectSimulationDataRcd, PPSDirectSimulationData);
 REGISTER_PLUGIN(PPSPixelTopologyRcd, PPSPixelTopology);
 REGISTER_PLUGIN(PPSAlignmentConfigRcd, PPSAlignmentConfig);
 REGISTER_PLUGIN(PPSAlignmentConfigurationRcd, PPSAlignmentConfiguration);
-REGISTER_PLUGIN(LHCInfoPerLSRcd, LHCInfoPerLS);
 
 REGISTER_PLUGIN_INIT(PPSAssociationCutsRcd, PPSAssociationCuts, InitAssociationCuts);
