+++ CondCore/RunInfoPlugins/src/plugin.cc
@@ -10,6 +10,8 @@
 #include "CondFormats/RunInfo/interface/FillInfo.h"
 #include "CondFormats/DataRecord/interface/LHCInfoRcd.h"
 #include "CondFormats/RunInfo/interface/LHCInfo.h"
+#include "CondFormats/DataRecord/interface/LHCInfoPerLSRcd.h"
+#include "CondFormats/RunInfo/interface/LHCInfoPerLS.h"
 
 REGISTER_PLUGIN(RunSummaryRcd, RunSummary);
 REGISTER_PLUGIN(RunInfoRcd, RunInfo);
@@ -17,3 +19,4 @@ REGISTER_PLUGIN(L1TriggerScalerRcd, L1TriggerScaler);
 REGISTER_PLUGIN(MixingRcd, MixingModuleConfig);
 REGISTER_PLUGIN(FillInfoRcd, FillInfo);
 REGISTER_PLUGIN(LHCInfoRcd, LHCInfo);
+REGISTER_PLUGIN(LHCInfoPerLSRcd, LHCInfoPerLS);
