+++ CondFormats/PPSObjects/src/classes_def.xml
@@ -104,6 +104,4 @@
     <class name="std::pair<unsigned int, PPSAlignmentConfiguration::SelectionRange>"/>
     <class name="PPSAlignmentConfiguration" class_version="0"/>
 
-    <class name="LHCInfoPerLS" class_version="0"/>
-
   </lcgdict>
