+++ CondFormats/PPSObjects/src/headers.h
@@ -18,5 +18,3 @@
 
 #include "CondFormats/PPSObjects/interface/PPSAlignmentConfig.h"
 #include "CondFormats/PPSObjects/interface/PPSAssociationCuts.h"
-
-#include "CondFormats/PPSObjects/interface/LHCInfoPerLS.h"
