--- CondFormats/PPSObjects/test/testSerializationLHCInfoPerLS.cc
@@ -1,8 +1,0 @@
-#include "CondFormats/Serialization/interface/Test.h"
-#include "CondFormats/PPSObjects/src/headers.h"
-
-int main() {
-  testSerialization<LHCInfoPerLS>();
-
-  return 0;
-}
