+++ CondFormats/RunInfo/src/T_EventSetup_LHCInfoPerLS.cc
@@ -1,4 +1,4 @@
-#include "CondFormats/PPSObjects/interface/LHCInfoPerLS.h"
+#include "CondFormats/RunInfo/interface/LHCInfoPerLS.h"
 #include "FWCore/Utilities/interface/typelookup.h"
 
 TYPELOOKUP_DATA_REG(LHCInfoPerLS);
