+++ CondFormats/RunInfo/src/headers.h
@@ -5,3 +5,4 @@
 #include "CondFormats/RunInfo/interface/MixingModuleConfig.h"
 #include "CondFormats/RunInfo/interface/FillInfo.h"
 #include "CondFormats/RunInfo/interface/LHCInfo.h"
+#include "CondFormats/RunInfo/interface/LHCInfoPerLS.h"
